requirements.txt
lxml
bs4
requests
sqlalchemy
simplejson
pyexecjs